Renaming of utility class objectstore.  Now called obcjectstorage.
